fpga: x411: added support for x411 (a ZCU111 based USRP) in the fpga code
The changes include following modifications of constraints in relation
to constraints for X410:
- creation of constraints with following changes:
 - commenting out all GPIO related pins as they are not crucial for the
  design to work and many of them (like gpios connected to
  daughter-boards) are not needed on ZCU111MGT_REFCLK_LMK0_N
 - assigning most of GPIOs that are available externally on X410 to
  external pins on ZCU111,
 - assigning clock inputs to the right pins,
 - commenting out connections to other devices that don't exist on
  ZCU111,
 - commenting out timing related pins (trigger, LMK sync) for now,
 - swapping QSFP0_* and QSFP1_* ports as only QSFP1 pins from X410 are
  connected to SFP+ cages on on ZCU111 (QSFP0 is connected to FMC port
  and requires additional extension card with SFP+ or QSFP ports),
 - commenting out timing constraints related to commented out pins.
 - addition of x411 dram pin constraints
 - PPS related inputs and outputs are mapped to following pins:
  - PPS_IN -> AMS_FPGA_REF_CLK (AP14) that is connected to J95
    SMA port
  - BASE_REFCLK_FPGA_P -> PMOD0_3_LS (H17),
    BASE_REFCLK_FPGA_N -> PMOD0_2_LS (H16)
  - PPS_LED -> GPIO_LED_0_LS (AR13)

Changes to ZynqMP configuration (this is not needed for synthesis but for
generation of configuration applied during early stage of boot):
 - creation of changed ZynqMP configuration for ZCU111 in
x411_ps_rfdc_bd.tcl ,
 - the main change is related to changing DDR from DDR2400 to DDR2166.
The changed of DDR clock results with changes for other clock that use
DDR clock (DPLL) as a reference,
 - adjusting assignment of ports (like UART) to the right (on ZCU111)
GPIOs,
TODO - check if ZCU's memory can work when configured to DDR2400).

Parts not working on X411 are disabled for X411 with defines.
Wires and dummy assignments are created in place of input and output
ports that don't exist on X411.

For ZCU111 (X411) additional outputs were added to the top module -
tx_enable lines that are connected to SFP+ modules associated with
QSFP0 port. They are set to 1 all the time.

Device tree files for x411 FPGA image were created.

Following changes were made to makefiles:
- addition of X411_X4_200 target in the main Makefile,
- addition of X411's FPGA chip model to PRODUCT_ID_MAP,