Changed ZynqMP configuration for ZCU111
The main change is related to changing DDR from
DDR2400 to DDR2166. The changed of DDR clock results with
changes for other clock that use DDR clock (DPLL)
as a reference (TODO - check if ZCU's memory can work when
configured to DDR2400).

Other changes include adjusting assignment of ports (like UART) to
the right (on ZCU111) GPIOs.

Connection for of processor GPIOs to the outside world was disabled.
On X410 they are used for stuff that doesn't exist on ZCU111.
This could be done just as well in x4xx.v and this is how it'll
be probably done in the future.

The CPLD jtag was also removed with its connection ports left
(a marginal optimization that I couldn't restrain myself from making ;P).