Commented out parts in top module
Some of the modules are not needed (like CPLD
or DB interfaces)

There are dummy wires and assignments created in
order to keep the top module synthesize correctly.
They'll need to be cleaned up in the future.

TODO: consider turning off these modules with a define
that is set when calling make.