ZCU111: changes that make life easier when working with NFS-based rootfs
Cmake reads SYSTEMD_SYSTEM_UNITDIR, SYSTEMD_SYSCTLDIR, UDEV_UDEVDIR
variables from PKG config. This is a problem when someone wants to install
UHD to a nfsroot directory. Some files won't be put there, but
they will be installed in the root path ('/') of the machine,
as if it was CMAKE_INSTALL_PREFIX.

The variables should be set manually as cmake parameters
like this:
cmake \
 -DSYSTEMD_SYSTEM_UNITDIR=/nfsroot/x4xx/lib/systemd/system
 -DSYSTEMD_SYSCTLDIR=/nfsroot/x4xx/usr/lib/sysctl.d \
 -DUDEV_UDEVDIR=/nfsroot/x4xx/lib/udev ...

Regarding usrp-adc... and usrp-hwd systemd files - I don't know
why @CMAKE_INSTALL_PREFIX@ was put there in the first place.
In my opinion this should be removed.