fpga: x411: change ADC tile for the second pair of channels
NI selected first and third ADC tiles. Each of them has 2 ADCs.
On ZCU111 they are connected to separate LMX2594 clock synthesizers.
Synchronizing LMX2594 output phases is not straightforward on ZCU111:
- https://e2e.ti.com/support/clock-timing-group/clock-and-timing/f/clock-timing-forum/1167523/lmx2594-can-t-get-lock-when-vco_phase_sync-is-enabled
- https://support.xilinx.com/s/question/0D54U00005p9au7SAA/how-to-get-phase-synced-between-lmx2594-chips-on-zcu111-eval-board?language=en_US

The current solution is to swap third tile for second tile, that is
connected to the same LMX2594 as the first tile. This way MTS is enough
to get their phases synchronized and have all RX channels phase
synchronization across power-cycles.

The selected solution is quite a nasty one and the root cause
is poor default naming of the x4xx_ps_rfdc_bd ports used by NI
(at least 'poor from my perspective'). The pragmatic decision is
to leave that naming alone, as currently changing it leads
to changes in too many places. Supporting these changes
would be a burden whenever related files are updated by NI.

The cost: it is necessary to remember that:
- port names starting from 'adc_tile226_' are really starting
from 'adc_tile225_',
- cAdc2xNcoUpdateBusy is really cAdc1xNcoUpdateBusy,
- cAdc2xNcoUpdateReq  is really cAdc1xNcoUpdateReq.